Refactoring tree structure, keep crashing
